requirements.txt
streamlit-back-camera-input
streamlit-camera-input-live